refactored results dataclasses
