+++ qc2/algorithms/base/__init__.py
@@ -1,10 +1,6 @@
 """Base Algo package."""
 from .base_algorithm import BaseAlgorithm
-from .base_algorithm_results import BaseAlgorithmResults
-from .vqe_base_results import VQEBASEResults
 
 __all__ = [
     "BaseAlgorithm",
-    "BaseAlgorithmResults",
-    "VQEBASEResults"
 ]
