--- qc2/algorithms/base/base_algorithm_results.py
@@ -1,6 +1,0 @@
-"""Module defining base class for all algorithms results."""
-from abc import ABC
-
-
-class BaseAlgorithmResults(ABC):
-    """Base class for all qc2 algorithms results."""
