+++ qc2/algorithms/pennylane/vqe/vqe.py
@@ -6,7 +6,7 @@ from pennylane import QNode
 from pennylane.operation import Operator
 from qc2.second_q.active_space import ActiveSpace
 from qc2.algorithms.base.base_algorithm import BaseAlgorithm
-from qc2.algorithms.algorithms_results import VQEResults
+from qc2.algorithms.results import VQEResults
 from qc2.algorithms.pennylane.ansatz.generate_ansatz import generate_ansatz
 from qc2.algorithms.pennylane.qubit_mappers.jordan_wigner import JordanWigner
 from qc2.qc2_driver import QC2
