+++ qc2/algorithms/qiskit/qpe/pebase.py
@@ -8,7 +8,7 @@ from qiskit.circuit.library import UnitaryGate
 from qc2.algorithms.base.base_algorithm import BaseAlgorithm
 from qc2.second_q.active_space import ActiveSpace
 
-from qc2.algorithms.algorithms_results import QPEResults
+from qc2.algorithms.results import QPEResults
 from qc2.algorithms.qiskit.qubit_mappers.jordan_wigner import JordanWigner
 from qc2.qc2_driver import QC2
 from qc2.second_q.active_space import ActiveSpace
